requirements.txt
pypwm
memorable_password
bottle
jinja2
cython==0.25.2
kivy
